Moved media controls to new file
